Pre: move algebrizer types to their own file.
